page folder created and Home, ProjectDetails components moved to page folder
