+++ src/App.css
@@ -766,10 +766,15 @@ button {
   font-weight: 700;
   display: flex;
   align-items: center;
-  gap: 14px;
+  /* gap: 14px; */
   padding: 15px 30px;
   background-color: #ffffff;
   max-height: 70px;
+}
+.skill-experience{
+  margin-right:-400px;
+  font-weight: 400;
+  
 }
 
 .skill-icon {
