+++ src/App.js
@@ -1,9 +1,9 @@
 import { Routes, Route } from "react-router-dom";
 import "./App.css";
 import Footer from "./components/Footer";
-import Home from "./components/Home";
+import Home from "./pages/Home";
 import Navigation from "./components/Navigation";
-import ProjectDetails from "./components/ProjectDetails";
+import ProjectDetails from "./pages/ProjectDetails";
 import NotFoundPage from "./components/NotFoundPage";
 
 function App() {
